IFS-2221-SP-Gen-BD-Approved-Or-NR
Implemented Duncan's comment to return BigDecimal from a native query
instead of an object.

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/competition/transactional/CompetitionServiceImpl.java
@@ -420,6 +420,6 @@ public class CompetitionServiceImpl extends BaseTransactionalService implements 
     @Override
     public ServiceResult<Long> countPendingSpendProfiles(Long competitionId) {
 
-        return serviceSuccess(((BigInteger)competitionRepository.countPendingSpendProfiles(competitionId)).longValue());
+        return serviceSuccess(competitionRepository.countPendingSpendProfiles(competitionId).longValue());
     }
 }
